refactor(review): extract VerdictBuilder from ReviewOrchestrator (#250)
Step 6 (the hotspot) of the decomposition: move the decision core into a
VerdictBuilder — buildResult (all overloads), tallyFindings/FindingTally,
the DiffStats record, toChangedFiles, and the check-run
conclusion/title/summary derivation.

Pure extract-class, no behaviour change. The APPROVE-gating guards move
VERBATIM: a review approves only with no outstanding new findings AND no
unresolved previous findings (#118/#130/#143 backstop merged into
previousStatuses), no offending CI checks (#217), and a non-truncated
diff (#234). summaryGenerator leaves the orchestrator (only buildResult
used it); followUpAnalyzer stays shared. review() now calls
verdictBuilder.buildResult / VerdictBuilder.{conclusion,title,summary}For
Result and VerdictBuilder.DiffStats. ReviewOrchestrator is down to ~475
LOC (from ~1600).

Tests: the orchestrator test builds a real VerdictBuilder from the mocked
summaryGenerator/followUpAnalyzer and the verdict cases (BuildResult,
ApproveBackstop, CiChecksGating, UnresolvedPreviousGating,
CheckRunPresentation, DiffStatsHelper) call it directly — pinning the
#211/#217/#234 guards. Full suite green, spotbugs + spotless clean.
